clean up test package structure
