--- alpha-core/src/test/java/at/ac/tuwien/kr/alpha/TestUtil.java
@@ -1,74 +1,0 @@
-/*
- * Copyright (c) 2019-2020 Siemens AG
- * All rights reserved.
- * 
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * 
- * 1) Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 
- * 2) Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package at.ac.tuwien.kr.alpha;
-
-import java.util.Collection;
-import java.util.stream.Collectors;
-
-import org.apache.commons.lang3.StringUtils;
-
-import at.ac.tuwien.kr.alpha.api.programs.atoms.Atom;
-import at.ac.tuwien.kr.alpha.api.terms.Term;
-import at.ac.tuwien.kr.alpha.commons.Predicates;
-import at.ac.tuwien.kr.alpha.commons.atoms.Atoms;
-import at.ac.tuwien.kr.alpha.commons.terms.Terms;
-import at.ac.tuwien.kr.alpha.core.common.AtomStore;
-import at.ac.tuwien.kr.alpha.core.common.NoGood;
-
-/**
- * Provides utility methods for test cases
- *
- */
-// TODO TestUtil/TestUtils?? Also, TestUtils duplicated
-// maybe introduce module "test-support" or smth
-public class TestUtil {
-
-	public static Atom atom(String predicateName, String... termStrings) {
-		Term[] terms = new Term[termStrings.length];
-		for (int i = 0; i < termStrings.length; i++) {
-			String termString = termStrings[i];
-			if (StringUtils.isAllUpperCase(termString.substring(0, 1))) {
-				terms[i] = Terms.newVariable(termString);
-			} else {
-				terms[i] = Terms.newConstant(termString);
-			}
-		}
-		return Atoms.newBasicAtom(Predicates.getPredicate(predicateName, terms.length), terms);
-	}
-
-	public static Atom atom(String predicateName, int... termInts) {
-		Term[] terms = new Term[termInts.length];
-		for (int i = 0; i < termInts.length; i++) {
-			terms[i] = Terms.newConstant(termInts[i]);
-		}
-		return Atoms.newBasicAtom(Predicates.getPredicate(predicateName, terms.length), terms);
-	}
-	
-	public static void printNoGoods(AtomStore atomStore, Collection<NoGood> noGoods) {
-		System.out.println(noGoods.stream().map(atomStore::noGoodToString).collect(Collectors.toSet()));
-	}
-
-}
