+++ alpha-core/src/test/java/at/ac/tuwien/kr/alpha/core/common/ProgramTest.java
@@ -23,7 +23,7 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package at.ac.tuwien.kr.alpha.common;
+package at.ac.tuwien.kr.alpha.core.common;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
