+++ alpha-core/src/test/java/at/ac/tuwien/kr/alpha/core/externals/AspStandardLibraryTest.java
@@ -1,4 +1,4 @@
-package at.ac.tuwien.kr.alpha.api.externals.stdlib;
+package at.ac.tuwien.kr.alpha.core.externals;
 
 import java.util.List;
 import java.util.Set;
@@ -8,7 +8,6 @@ import org.junit.Test;
 
 import at.ac.tuwien.kr.alpha.api.terms.ConstantTerm;
 import at.ac.tuwien.kr.alpha.commons.terms.Terms;
-import at.ac.tuwien.kr.alpha.core.externals.AspStandardLibrary;
 
 public class AspStandardLibraryTest {
