+++ alpha-core/src/test/java/at/ac/tuwien/kr/alpha/core/grounder/ChoiceGrounder.java
@@ -27,10 +27,10 @@
  */
 package at.ac.tuwien.kr.alpha.core.grounder;
 
-import static at.ac.tuwien.kr.alpha.common.NoGoodTest.fromOldLiterals;
 import static at.ac.tuwien.kr.alpha.commons.util.Util.entriesToMap;
 import static at.ac.tuwien.kr.alpha.commons.util.Util.entry;
 import static at.ac.tuwien.kr.alpha.core.common.NoGood.headFirst;
+import static at.ac.tuwien.kr.alpha.core.common.NoGoodTest.fromOldLiterals;
 import static java.util.Arrays.asList;
 import static java.util.Collections.singleton;
