+++ alpha-core/src/test/java/at/ac/tuwien/kr/alpha/core/grounder/heuristics/GrounderHeuristicConfigurationTest.java
@@ -23,7 +23,7 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package at.ac.tuwien.kr.alpha.grounder.heuristics;
+package at.ac.tuwien.kr.alpha.core.grounder.heuristics;
 
 
 import static at.ac.tuwien.kr.alpha.api.config.GrounderHeuristicsConfiguration.PERMISSIVE_STRING;
