+++ alpha-core/src/test/java/at/ac/tuwien/kr/alpha/core/test/util/DependencyGraphUtils.java
@@ -1,4 +1,4 @@
-package at.ac.tuwien.kr.alpha.test.util;
+package at.ac.tuwien.kr.alpha.core.test.util;
 
 import java.util.ArrayList;
 import java.util.HashSet;
